backend/requirements.txt
# A股盘后静态分析系统依赖

# 数据库
psycopg2-binary==2.9.9

# 数据源
baostock==0.8.9

# 配置管理
PyYAML==6.0.1

# 数据处理
pandas==2.1.4

# 命令行工具
click==8.1.7

# 日期时间处理
python-dateutil==2.8.2

# 拼音处理（文档要求）
pypinyin==0.51.0

# 工具库
tqdm==4.66.1